requirements.txt
django==4.1
django-cors-headers==4.2
djangorestframework==3.14.0
drf-yasg==1.21.7
pillow==10.3
psycopg2-binary==2.9.9
python-decouple==3.8
